Segmenting and Clustering Neighborhoods in Toronto
This is part of Week 3's assignment. There are three parts to this assignment that I have completed in one notebook: 1) Scraping the following Wikipedia page, https://en.wikipedia.org/wiki/List_of_postal_codes_of_Canada:_M, in order to obtain the data that is in the table of postal codes and to transform the data into a pandas dataframe 2) Obtaining the latitude and the longitude coordinates of each neighborhood 3) Explore and cluster the neighborhoods in Toronto.